Minor Modifications on A-Star Utility Methods

+++ Pathfinder/Pathfinder/a_star_search.cpp
@@ -61,7 +61,7 @@ void Init_Grid()
 
 float Distance(Node* a, Node* b)
 {
-	return sqrtf((a->x - b->x) * (a->x - b->x) + (a->y - b->y) * (a->y - b->y));
+	return sqrtf(((float)a->x - b->x) * (a->x - b->x) + (a->y - b->y) * (a->y - b->y));
 }
 
 float Heuristic(Node* a, Node* b)
